+++ src/Core/Api/V1/Controllers/usersController.cs
@@ -1,4 +1,5 @@
 using e_commerce_server.src.Core.Database;
+using e_commerce_server.src.Core.Modules.Auth.Dto;
 using e_commerce_server.src.Core.Modules.Review.Service;
 using e_commerce_server.src.Core.Modules.User.Dto;
 using e_commerce_server.src.Core.Modules.User.Service;
@@ -126,6 +127,22 @@ namespace e_commerce_server.src.Core.Api.V1.Controllers
             catch (HttpException ex)
             {
                 return StatusCode((int)ex.StatusCode, ex.Response);
+            }
+        }
+
+        [HttpPost("change-password")]
+        [Authorize]
+        public IActionResult ChangePassword(ChangePasswordDto model)
+        {
+            try
+            {
+                var idClaim = HttpContext.User.FindFirst("id")?.Value;
+
+                return Ok(userService.ChangePassword(model, Convert.ToInt32(idClaim)));
+            }
+            catch (HttpException ex)
+            {
+                return StatusCode((int) ex.StatusCode, ex.Response);
             }
         }
     }
