+++ src/Core/Modules/Auth/Service/AuthEnum.cs
@@ -5,7 +5,8 @@
         public const string LOGIN_SUCCESS = "Login successfully";
         public const string LOGIN_INCORRECT = "Email or password is incorrect";
         public const string DUPLICATE_EMAIL = "This email is already exists";
-        public const string PASSWORDS_NOT_MATCH = "Password and confirmation password does not match";
+        public const string CONFIRM_PASSWORDS_NOT_MATCH = "Password and confirmation password is not match";
+        public const string PASSWORDS_NOT_MATCH = "Password is not correct";
         public const string DUPLICATE_PHONE_NUMBER = "This phone number is already exists";
         public const string REGISTER_SUCCESS = "Register successfully";
         public const string INVALID_REFRESH_TOKEN = "Invalid refresh token";
