+++ src/Core/Modules/Auth/Service/AuthService.cs
@@ -55,7 +55,7 @@ namespace e_commerce_server.src.Core.Modules.Auth.Service
         {
             if (model.password != model.confirm_password)
             {
-                throw new BadRequestException(AuthEnum.PASSWORDS_NOT_MATCH);
+                throw new BadRequestException(AuthEnum.CONFIRM_PASSWORDS_NOT_MATCH);
             }
 
             var existingUser = userRepository.GetUserByEmail(model.email);
@@ -150,7 +150,7 @@ namespace e_commerce_server.src.Core.Modules.Auth.Service
         {
             if (model.password != model.confirm_password)
             {
-                throw new BadRequestException(AuthEnum.PASSWORDS_NOT_MATCH);
+                throw new BadRequestException(AuthEnum.CONFIRM_PASSWORDS_NOT_MATCH);
             }
 
             var user = Optional.Of(userRepository.GetUserByResetToken(model.reset_token)).ThrowIfNotPresent(new BadRequestException(AuthEnum.INVALID_TOKEN)).Get();
@@ -171,31 +171,6 @@ namespace e_commerce_server.src.Core.Modules.Auth.Service
                 };
             }
             throw new BadRequestException(AuthEnum.EXPIRED_TOKEN);
-        }
-        public object ChangePassword(ChangePasswordDto model, int id)
-        {
-            var user = userRepository.GetUserById(id);
-
-            if (!bCryptService.Verify(model.password, user.password))
-            {
-                throw new BadRequestException(AuthEnum.PASSWORDS_NOT_MATCH);
-            }
-            else
-            {
-                if (model.newpassword != model.confirmnewpassword)
-                {
-                    throw new BadRequestException(AuthEnum.PASSWORDS_NOT_MATCH);
-                }
-                user.update_at = DateTime.Now;
-                user.password = bCryptService.Hash(model.newpassword);
-            }
-            
-            userRepository.CreateOrUpdateUser(user);
-
-            return new
-            {
-                message = AuthEnum.UPDATE_PASSWORD_SUCCESS
-            };
         }
     }
 }
